requirements.txt
Flask==2.3.3
transformers==4.34.0
Pillow==10.0.1
torch==2.0.1 
